Add RSpec
To perform deeper refactoring a unit test suite will be useful.

I like to make explicit the distinction between the specs which document
the public API (tagged as :public) and those which are only refactoring
guards and can be changed when necessary (tagged as :protected or
:private depending on how stable they are intended to be).

The 'spec:public' and 'spec:development' Rake tasks reflect
that distinction.