Refactor move hooks registration to Runner
Minor use parens in multiple arguments method calls
Minor use parens in non-tivial method calls

The DreddHooks::Methods exposes the public DSL of DreddHooks.
As an interface, I would like it to be as simple as possible,
and as immutable as possible.

Since the DSL and the runner were closely related already, the latter
seems to be a good place to store the hooks that need to be runned.

Also, because the DSL is an interface for the runner, I prefer the
dependency to go from the interface to the runner, that the contrary.

Finally, I made the runner a Singleton because it was already unique
when defined as a Module, and because it make trivial its retrieval
from the DSL.