Refactor move Server::EVENTS definition
When adding support for new hooks, both the hooks definition and
the events definition needs to be updated. It makes sense to me to
keep them in the same file.

As the events definition refers to the hooks names, in order to make
maintenance easier, I think it also makes sense to check consistency,
and break the test suite when required definitions are missing.