Refactor encapsulate DreddHooks::CLI
This refactoring should allow to start using Thor to build the CLI with
minimal disruption.

Using Thor would change the API (starting the server would require to
dredd-hooks-ruby start instead of dredd-hooks-ruby) but would allow to
build a properly documented CLI (dredd-hooks-ruby help and so on...)
with minimal effort.

See http://whatisthor.com