Temporary use local copy of Aruba
Aruba failures for versions >= 0.7.0 seem to be fixed by removing
the @debug Cucumber tags. I am not sure what is the expected behaviour
of these tags, but Aruba did undergo quite a few changes between
v0.6.2 and v0.7.0 (refactoring and features).

Using a local copy of Aruba allows to use 'git bisect' to discover
exactly what change in Aruba started causing failures.
From there is should be easier to figure out if the failure is
related of a change in the @debug behaviour, a mistake in how
we use it, or an Aruba bug (unlikely, but still possible).

See https://github.com/apiaryio/dredd-hooks-ruby/issues/26

Usage:

    # Clone Aruba in ../aruba
    git clone git@github.com:cucumber/aruba.git ../aruba

    # In order to test a specific version of Aruba, use tags:
    cd ../aruba
    git checkout v0.6.2

    cd ../dredd-hooks-ruby
    bundle install && rake cucumber # is 'good' (green) with v0.6.2

    # Speed up the discovery with 'git bisect':
    cd ../aruba
    git bisect start
    git bisect bad v0.7.0
    git bisect good v0.6.2

    # Repeat:
    cd ../dredd-hooks-ruby
    bundle install && rake cucumber # the features do pass or fail
    cd ../aruba
    git bisect good # resp. 'bad' if the DreddHooks features did fail

    # Until bisect stops pointing at the suspicious commit.
    git bisect reset # stop bisecting and restores everthing as it was